Rename BookingsPage.js to MyBookingsPage.js, Update bookingsApiSlice
